refactor(team): split team.rs + store.rs to satisfy file-size gate (#665)
CI's check-file-size.sh enforces 400 LOC; the initial #665 commit pushed
src/commands/team.rs to 950 LOC and src/state/store.rs to 450 LOC.

Splits + dedup (also addresses simplify-pass HIGH findings):

- src/commands/team.rs (950 → 383):
  - Extract launch wiring (LaunchOpts, LaunchOutcome, SchedulerRunner trait,
    ProductionSchedulerRunner, build_metas_from_args, launch_headless,
    print_launch_outcome) into src/commands/team_launch.rs
  - Extract test module into src/commands/team_tests.rs (sidecar via
    #[path], same pattern as commands/init.rs ↔ init_tests.rs)
  - Remove duplicate validate_preset_name (loader's stricter version is
    now the single source of truth — pub(crate) in loader)
  - Remove duplicate write_preset_file (use loader's pub(crate) version)
  - Remove primitive_label helper — use existing Primitive::label() const fn

- src/state/store.rs (450 → 124):
  - Extract test module into src/state/store_tests.rs
  - Replace inline legacy_json blob in load_legacy_state_without_handoff_fields_succeeds
    with include_str!("../../tests/fixtures/state/v0.json") — same fixture,
    one source of truth

No behavior change. 4526 tests still pass.
